refact : ncp 설정 관련 값들을 Ncp Properties로 변경 및 리펙토링
ncp properties(component)에서 ncp에 필요한 값을 받아서 처리
